+++ libMx/mx_digital_output.h
@@ -7,7 +7,7 @@
  *
  *--------------------------------------------------------------------------
  *
- * Copyright 1999, 2001, 2006 Illinois Institute of Technology
+ * Copyright 1999, 2001 Illinois Institute of Technology
  *
  * See the file "LICENSE" for information on usage and redistribution
  * of this file, and for a DISCLAIMER OF ALL WARRANTIES.
@@ -23,7 +23,7 @@ typedef struct {
 	MX_RECORD *record; /* Pointer to the MX_RECORD structure that points
                             * to this device.
                             */
-	uint32_t value;
+	unsigned long value;
 } MX_DIGITAL_OUTPUT;
 
 #define MXLV_DOU_VALUE 5001
@@ -44,11 +44,11 @@ typedef struct {
 	mx_status_type ( *write ) ( MX_DIGITAL_OUTPUT *doutput );
 } MX_DIGITAL_OUTPUT_FUNCTION_LIST;
 
-MX_API mx_status_type mx_digital_output_read( MX_RECORD *doutput,
-						uint32_t *value );
+MX_API mx_status_type mx_digital_output_read(MX_RECORD *doutput,
+							unsigned long *value);
 
-MX_API mx_status_type mx_digital_output_write( MX_RECORD *doutput,
-						uint32_t value );
+MX_API mx_status_type mx_digital_output_write(MX_RECORD *doutput,
+							unsigned long value);
 
 extern MX_RECORD_FUNCTION_LIST mx_digital_output_record_function_list;
